@@ -29,47 +29,49 @@
     }
   },
   "dependencies": {
-    "@webcomponents/webcomponentsjs": "^2.1.3",
-    "enzyme": "^3.4.4",
-    "enzyme-adapter-react-16": "^1.2.0",
+    "@webcomponents/webcomponentsjs": "^2.3.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
     "immutable": "^4.0.0-rc.9",
-    "jest": "^22.4.3",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "react-redux": "^5.0.7",
-    "react-web-component": "^1.3.0",
-    "reactive-elements": "^0.10.0",
-    "reduce-reducers": "^0.1.2",
-    "redux": "^3.7.2",
-    "reselect": "^3.0.1"
+    "jest": "^24.9.0",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "react-helmet": "^5.2.1",
+    "react-redux": "^7.1.3",
+    "react-web-component": "^2.0.1",
+    "reduce-reducers": "^1.0.4",
+    "redux": "^4.0.4",
+    "reselect": "^4.0.0",
+    "styled-components": "^4.4.1"
   },
   "devDependencies": {
-    "@types/react": "^16.4.11",
-    "@types/react-dom": "^16.0.7",
-    "@types/enzyme": "^3.1.13",
-    "@types/enzyme-adapter-react-16": "^1.0.3",
-    "@types/jest": "^22.2.2",
-    "@types/react-redux": "^6.0.6",
+    "@types/enzyme": "^3.10.3",
+    "@types/enzyme-adapter-react-16": "^1.0.5",
+    "@types/jest": "^24.0.22",
+    "@types/react": "^16.9.11",
+    "@types/react-dom": "^16.9.4",
+    "@types/react-redux": "^7.1.5",
+    "@types/styled-components": "^4.4.0",
     "bundle-loader": "^0.5.6",
-    "cross-env": "^5.2.0",
-    "css-loader": "^1.0.0",
+    "cross-env": "^6.0.3",
+    "css-loader": "^3.2.0",
     "html-webpack-plugin": "^3.2.0",
-    "node-sass": "^4.9.3",
-    "sass-loader": "^7.1.0",
+    "node-sass": "^4.13.0",
     "source-map-loader": "^0.2.4",
-    "style-loader": "^0.22.1",
-    "ts-jest": "^22.4.2",
+    "style-loader": "^1.0.0",
+    "ts-jest": "^24.1.0",
     "ts-lint": "^4.5.1",
-    "ts-loader": "^4.1.0",
-    "tslint": "^5.11.0",
-    "tslint-react": "^3.6.0",
-    "typescript": "^2.8.1",
-    "webpack": "^4.17.1",
-    "webpack-cli": "^2.0.14",
-    "webpack-dev-server": "^3.1.5"
+    "ts-loader": "^6.2.1",
+    "tslint": "^5.20.1",
+    "tslint-config-prettier": "^1.18.0",
+    "tslint-react": "^4.1.0",
+    "typescript": "^3.7.2",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0"
   },
   "scripts": {
-    "start:dev": "cross-env NODE_ENV=development webpack-dev-server",
+    "start:dev": "cross-env NODE_ENV=development webpack-dev-server --hot",
     "test": "jest",
     "build": "cross-env NODE_ENV=production webpack -p",
     "lint": "tslint -c tslint.json 'src/**/*.{ts,tsx} --fix"
