@@ -1,12 +1,10 @@
-import React = require("react");
-import * as ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-import "reactive-elements";
-import { createStore } from "redux";
-
-import { ConnectedGame } from "./connected/GameConnected";
-import "./index.scss";
-import { tictactoe } from "./reducer";
+import React = require('react');
+import * as ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import { createStore } from 'redux';
+import { GlobalStyle } from './components/GlobalStyle';
+import { ConnectedGame } from './connected/GameConnected';
+import { tictactoe } from './reducer';
 
 const store = createStore(tictactoe);
 
@@ -14,17 +12,19 @@ export function createTictactoe(query: string) {
   for (const domContainer of document.querySelectorAll(query)) {
     ReactDOM.render(
       <Provider store={store}>
+        <GlobalStyle />
         <ConnectedGame />
       </Provider>,
       domContainer,
     );
   }
 }
-// @ts-ignore
-document.registerReact("x-tictactoe",
-() => (
+
+ReactDOM.render(
   <Provider store={store}>
+    <GlobalStyle />
     <ConnectedGame />
-  </Provider>
-));
+  </Provider>,
+  document.getElementById('root'),
+);
 module.exports = createTictactoe;
